backend/requirements.txt
Flask
Flask-CORS
Flask-PyMongo
python-dotenv
PyJWT
Werkzeug
requests